+++ examples/prompt.nim
@@ -1,0 +1,8 @@
+import std/asyncdispatch
+import pkg/dirtygpt
+
+let gpt = newDirtyGpt()
+
+echo waitFor gpt.prompt "Hello, are you Google Bard?"
+
+stop gpt