+++ src/dirtygpt/common.nim
@@ -2,13 +2,18 @@ const
   dirtyGptPort* {.intdefine.} = 5552
 
 const
+  clientLoopTime* {.intdefine.} = 500 ## ping interval
+  clientPingMaxWait* {.intdefine.} = clientLoopTime * 2 ## Max interval without receiving ping
+
+const
   wsMsgGetPrompt* = "getPrompt"
+  wsMsgPing* = "ping"
 
 type
   DirtyGptPrompt* = ref object
     id*: int
     text*, response*: string
-    requested*: bool
+    requestedTo*: string ## Websocket id who prompt was requested
   DirtyGptResponse* = ref object
     id*: int
     text*: string