+++ src/dirtygpt/header.nim
@@ -9,9 +9,6 @@ const userscriptHeader* = genMetadataBlock(
     "https://chat.openai.com/*"
   ],
   runAt = GmRunAt.docIdle,
-  grant = [
-    GmPermitions.xmlHttpRequest,
-  ],
   require = [
     "https://unpkg.com/turndown/dist/turndown.js",
   ]